move extension methods to data layer
